Split out test code using tsc --build
This makes it easier to package this library, as we can avoid generating
test output altogether.  We also gain incremental builds.

Generating map files for type definitions hopefully helps IDE navigation
for consumers.
