+++ .gitignore
@@ -1,13 +1,11 @@
 *.log
 .DS_Store
-node_modules
-dist
-.idea
-.~
+/node_modules/
+/dist/
+/build/
+/.idea/
+*~
 .*.swp
 .parcel-cache
 .tool-versions
 .npmrc
-
-*~
-.*.swp
