--- .npmignore
@@ -1,6 +1,0 @@
-*.test.*
-test
-.idea/
-.github/
-*~
-!src/*
