+++ .pre-commit-config.yaml
@@ -28,7 +28,3 @@ repos:
     rev: v3.0.0-alpha.0
     hooks:
       - id: prettier
-#- repo: https://github.com/andrewaylett/pre-commit-hooks
-#  rev: v1.0      # or specific tag
-#  hooks:
-#  - id: npm-test
