+++ package.json
@@ -14,10 +14,10 @@
   "types": "dist/index.d.ts",
   "scripts": {
     "build": "npm run buildonly",
-    "buildonly": "tsc",
+    "buildonly": "tsc -b --verbose .",
     "test": "NODE_OPTIONS=--experimental-vm-modules jest --testPathIgnorePatterns='<rootDir>/(?:.+?)/postbuild.*|.*\\.js'",
-    "postbuild": "NODE_OPTIONS=--experimental-vm-modules jest src/postbuild.test.ts",
-    "lint": "eslint src --ext .ts",
+    "postbuild": "NODE_OPTIONS=--experimental-vm-modules jest test/postbuild.test.ts",
+    "lint": "eslint src test --ext .ts",
     "prebuild": "npm run test",
     "pretest": "npm run lint",
     "lint:fix": "npm run lint -- --fix",
@@ -39,7 +39,7 @@
     "eslint": "8.23.0",
     "eslint-config-prettier": "8.5.0",
     "eslint-plugin-prettier": "4.2.1",
-    "extend-expect": "0.2.0",
+    "extend-expect": "1.0.1",
     "jest": "28.1.3",
     "prettier": "2.7.1",
     "semver": "7.3.7",
