+++ src/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "..",
+  "include": ["./**/*.ts"],
+  "compilerOptions": {
+    "outDir": "../dist",
+    "tsBuildInfoFile": "../build/src.tsbuildinfo"
+  }
+}
