+++ test/gcmap.test.ts
@@ -16,8 +16,8 @@
 
 import { describe, it } from '@jest/globals';
 
-import { expect } from './test/expect';
-import { gc, sortByKey } from './gcmap';
+import { expect } from './expect';
+import { gc, sortByKey } from '../src/gcmap';
 
 describe('Sort by key', () => {
     it('retains an identity sort', () => {
