+++ test/tsconfig.json
@@ -1,0 +1,13 @@
+{
+  "extends": "..",
+  "include": ["./**/*.ts"],
+  "references": [
+    {
+      "path": "../src"
+    }
+  ],
+  "compilerOptions": {
+    "noEmit": true,
+    "tsBuildInfoFile": "../build/test.tsbuildinfo"
+  }
+}
